Fix rotted devman backend to be compatible with the config service.
Ready to be hooked up on the front end... Ex, query like this:
http://transient.syncfs.com:9020/devices/4afba56666f341d4b15fb56ab1cb9ede

+++ src/devman/src/com/aerofs/devman/server/DeviceManagementService.java
@@ -45,7 +45,7 @@ public final class DeviceManagementService extends Service<DeviceManagementServi
                 configuration.getRedisConfiguration().getPort());
         provider.init_(
                 configuration.getRedisConfiguration().getHost(),
-                configuration.getRedisConfiguration().getPort());
+                (short) configuration.getRedisConfiguration().getPort());
 
         l.warn("verkehr host=" + configuration.getVerkehrConfiguration().getHost() + " port=" +
                 configuration.getVerkehrConfiguration().getPort());
